comments fix
