+++ cmd/main.go
@@ -2,6 +2,7 @@ package main
 
 import (
 	"log/slog"
+	"os"
 
 	"github.com/logmanager-oss/dashboards-migrator/cmd/cli"
 )
@@ -10,5 +11,6 @@ func main() {
 	err := cli.CLIStart()
 	if err != nil {
 		slog.Error(err.Error())
+		os.Exit(1)
 	}
 }
